renamed args folder to singular arg
