+++ backend/src/graphql/arg/SubscribeNewsletterArgs.ts
