+++ backend/src/graphql/resolver/UserResolver.ts
@@ -9,11 +9,11 @@ import { SendPasswordResetEmailResponse } from '../model/SendPasswordResetEmailR
 import { UpdateUserInfosResponse } from '../model/UpdateUserInfosResponse'
 import { User } from '../model/User'
 import encode from '../../jwt/encode'
-import ChangePasswordArgs from '../args/ChangePasswordArgs'
-import CheckUsernameArgs from '../args/CheckUsernameArgs'
-import CreateUserArgs from '../args/CreateUserArgs'
-import UnsecureLoginArgs from '../args/UnsecureLoginArgs'
-import UpdateUserInfosArgs from '../args/UpdateUserInfosArgs'
+import ChangePasswordArgs from '../arg/ChangePasswordArgs'
+import CheckUsernameArgs from '../arg/CheckUsernameArgs'
+import CreateUserArgs from '../arg/CreateUserArgs'
+import UnsecureLoginArgs from '../arg/UnsecureLoginArgs'
+import UpdateUserInfosArgs from '../arg/UpdateUserInfosArgs'
 import { apiPost, apiGet } from '../../apis/HttpRequest'
 import {
   klicktippRegistrationMiddleware,
